+++ typescript/data/incorrect-input.txt
@@ -1,0 +1,7 @@
+
+1 1 E
+RFRFRFRF
+3 2 N
+FRRFLLFFRRFLL
+0 3 W
+LLFFFLFLFL
